restructured file system and changed simulator
