--- CoT_Stream_Simulator/README.md
@@ -1,5 +1,0 @@
-# CoT Stream Simulator
-
-## To run:
-run server.py \
-run simulated_client.py
