--- CoT_Stream_Simulator/__init__.py
@@ -1,1 +1,0 @@
-from ..SITCoTmsg.SITCoTmsg import *
