--- SITCoTmsg/SITCoTmsg/__init__.py
@@ -1,1 +1,0 @@
-from .SITCoTmsg import *
